// src/react/redux/types/ui/dashboardVisualization.ts
export const DataVisualizationsList = [TableKey, TimeSeriesKey, HistogramKey, Histogram3DKey] as const;
export type DataVisTypes = (typeof DataVisualizationsList)[number];

const VisPropertiesBase = <T extends VisTypes>(type: T) => z.object({ type: z.literal(type) });

const DataVisPropertiesBase = <T extends DataVisTypes>(type: T) =>
  VisPropertiesBase(type).extend({ connection: z.string() });

export const VisPropertiesMap = {
  [TextAreaKey]: VisPropertiesBase(TextAreaKey).extend({
    text: z.string().nullable()
  }),
  [TableKey]: DataVisPropertiesBase(TableKey).extend({
    rowsperpage: z.number()
  }),
  [TimeSeriesKey]: DataVisPropertiesBase(TimeSeriesKey).extend({
    numtraces: z.number()
  }),
  [HistogramKey]: DataVisPropertiesBase(HistogramKey).extend({}),
  [Histogram3DKey]: DataVisPropertiesBase(Histogram3DKey).extend({})
};
// ...
type VisDataSetDataMap = {
  [TableKey]: {
    $format: "csv";
    data: string;
  };
  [TimeSeriesKey]: {
    index: {
      name: string;
      units: string;
      type: "Double" | "Integer";
      dataBase64: string;
    };
    columns: {
      name: string;
      units: string;
      type: "Double" | "Integer";
      dataBase64: string;
    }[];
  };
  [HistogramKey]: {
    index: {
      name: string;
      units: string;
      type: "Double" | "Integer";
      dataBase64: string;
    };
    histograms: {
      chanName: string;
      type: "h2axis";
      data: {
        x: {
          start: string;
          midPoint: string;
          end: string;
          label: string;
        }[];
        xUnits: string;
        xLabel: string;
        y: {
          value: string;
        }[];
        yUnits: string;
        yLabel: string;
      };
    }[];
  };
  [Histogram3DKey]: {
    index: {
      name: string;
      units: string;
      type: "Double" | "Integer";
      dataBase64: string;
    };
    histograms: {
      chanName: string;
      type: "h3axis";
      data: {
        x: {
          start: string;
          midPoint: string;
          end: string;
          label: string;
        }[];
        xUnits: string;
        xLabel: string;
        y: {
          start: string;
          midPoint: string;
          end: string;
          label: string;
        }[];
        yUnits: string;
        yLabel: string;
        z: {
          value: string;
        }[];
        zUnits: string;
        zLabel: string;
      };
    }[];
  };
};
// ...
export type StatelessVisDataSetType<T extends DataVisTypes> = {
  type: T;
  state: "Loaded";
  index: {
    name: string;
    type: "Double" | "String" | "Integer" | "Histogram";
    units: string;
  };
  dataColumns: Array<{
    name: string;
    type: "Double" | "String" | "Integer" | "Histogram";
    units: string;
  }>;
};

type VisDataSetType<T extends DataVisTypes> =
  | StatelessVisDataSetType<T>
  | {
      type: T;
      state: "Loading";
    };

type VisPropertiesType<T extends VisTypes> = z.infer<(typeof VisPropertiesMap)[T]>;

type VisDataRefreshParamsType<T extends DataVisTypes> = VisDataRefreshParamsMap[T];

type VisDataSetZodType<T> = T extends DataVisTypes ? ZodType<VisDataSetType<T>, ZodTypeDef, VisDataSetType<T>> : never;

export type DataSet<T extends DataVisTypes> = VisDataSetType<T>;
// ...
export type VisDataSetDataType<T extends DataVisTypes> = VisDataSetDataMap[T];
// ...
export type FilteredVisualizationProperties<T extends VisTypes> = z.infer<(typeof VisPropertiesMap)[T]>;
// ...
export type DataSetId = { dataSetId: string };

export type DataSetDetails = {
  dashboardId: string;
  visualizationId: string;
} & DataSetId;